script/lakesoul-e2e/pyproject.toml
[project]
name = "e2e"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kubernetes>=33.1.0",
    "opendal>=0.46.0",
    "attrs>=25.3.0",
    "pathspec>=0.12.1",
    "setproctitle>=1.3.6",
    "annotated-types>=0.7.0",
    "colorlog>=6.9.0",
    "docker>=7.1.0",
    "gitpython>=3.1.45",
    "tqdm>=4.67.1",
    "faker>=37.5.3",
    "pyarrow>=21.0.0",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "jinja2>=3.1.6",
    "psutil>=7.0.0",
]

[project.scripts]
e2etest = "e2e:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
